Fix possible loss of data warning

+++ src/util/piped_process.cpp
@@ -345,7 +345,7 @@ piped_processt::send_responset piped_processt::send(const std::string &message)
     return send_responset::ERRORED;
   }
 #ifdef _WIN32
-  if(!WriteFile(child_std_IN_Wr, message.c_str(), message.size(), NULL, NULL))
+  if(!WriteFile(child_std_IN_Wr, message.c_str(), narrow<DWORD>(message.size()), NULL, NULL))
   {
     // Error handling with GetLastError ?
     return send_responset::FAILED;
